# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CtxOS API
  description: Context-driven OS Intelligence and Analysis API
  version: 1.0.0
  contact:
    name: CtxOS Team
    email: api@ctxos.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.ctxos.com
    description: Production server
...
paths:
...
  /api/v1/score:
    post:
      tags: [scoring]
      summary: Score entity
      description: Score an entity using specified scoring engines
      operationId: scoreEntity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreRequest'
      responses:
        '200':
          description: Entity scored successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoringResult'
        '401':
          description: Unauthorized
        '403':
          description: Insufficient permissions
        '422':
          description: Validation error
  /api/v1/score/batch:
    post:
      tags: [scoring]
      summary: Batch score entities
      description: Score multiple entities in parallel
      operationId: batchScore
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchScoreRequest'
      responses:
        '200':
          description: Batch scoring completed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoringResult'
        '400':
          description: Too many entities in batch
  /api/v1/score/history/{entity_id}:
    post:
      tags: [scoring]
      summary: Get scoring history
      description: Get historical scoring data for an entity
      operationId: getScoringHistory
      security:
        - bearerAuth: []
      parameters:
        - name: entity_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HistoryRequest'
      responses:
        '200':
          description: Historical data retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /api/v1/score/engines:
    get:
      tags: [scoring]
      summary: List scoring engines
      description: Get list of available scoring engines
      operationId: listScoringEngines
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Engines list retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnginesList'
  /api/v1/score/aggregate:
    post:
      tags: [scoring]
      summary: Get aggregate scores
      description: Get aggregate scoring statistics for multiple entities
      operationId: getAggregateScores
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AggregateRequest'
      responses:
        '200':
          description: Aggregate scores retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AggregateResponse'
  /api/v1/score/compare:
    post:
      tags: [scoring]
      summary: Compare entities
      description: Compare scores between multiple entities
      operationId: compareEntities
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompareRequest'
      responses:
        '200':
          description: Comparison completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompareResponse'
  /api/v1/score/status:
    get:
      tags: [scoring]
      summary: Scoring service health
      description: Get scoring service health status
      operationId: scoringHealth
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Health status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    HealthStatus:
      type: object
      required: [service, status, timestamp]
      properties:
        service:
          type: string
          description: Service name
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
          description: Service health status
        timestamp:
          type: string
          format: date-time
          description: Health check timestamp
        features:
          type: object
          description: Service features
        statistics:
          type: object
          description: Service statistics
    PaginatedResponse:
      type: object
      required: [items, total, limit, offset, has_next]
      properties:
        items:
          type: array
          items:
            type: object
          description: Response items
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
          description: Items per page
        offset:
          type: integer
          description: Number of items skipped
        has_next:
          type: boolean
          description: Whether there are more items

    # Data Model Schemas
    EntityType:
      type: string
      enum: [host, domain, ip, user, service, application, network, database]
    SignalType:
      type: string
      enum: [vulnerability, port, service, dns, ssl_certificate, configuration, activity, authentication, dependency, subdomain, whois]
    SeverityLevel:
      type: string
      enum: [critical, high, medium, low, info]
    Context:
      type: object
      required: [entity, signals]
      properties:
        entity:
          $ref: '#/components/schemas/Entity'
        signals:
          type: array
          items:
            $ref: '#/components/schemas/Signal'
    Entity:
      type: object
      required: [id, entity_type, name]
      properties:
        id:
          type: string
          description: Entity ID
        entity_type:
          $ref: '#/components/schemas/EntityType'
          description: Entity type
        name:
          type: string
          description: Entity name
        description:
          type: string
          description: Entity description
        properties:
          type: object
          description: Entity properties
    Signal:
      type: object
      required: [id, source, signal_type, severity, description]
      properties:
        id:
          type: string
          description: Signal ID
        source:
          type: string
          description: Signal source
        signal_type:
          $ref: '#/components/schemas/SignalType'
          description: Signal type
        severity:
          $ref: '#/components/schemas/SeverityLevel'
          description: Signal severity
        description:
          type: string
          description: Signal description
        timestamp:
          type: string
          format: date-time
          description: Signal timestamp
        entity_id:
          type: string
          description: Associated entity ID
        properties:
          type: object
          description: Signal properties

    # Scoring Schemas
    ScoreRequest:
      type: object
      required: [context, engines]
      properties:
        context:
          $ref: '#/components/schemas/Context'
          description: Entity context
        engines:
          type: array
          items:
            type: string
          description: Scoring engines to use
        include_recommendations:
          type: boolean
          default: true
          description: Include recommendations
    BatchScoreRequest:
      type: object
      required: [contexts, engines]
      properties:
        contexts:
          type: array
          items:
            $ref: '#/components/schemas/Context'
          description: Entity contexts
        engines:
          type: array
          items:
            type: string
          description: Scoring engines to use
        parallel:
          type: boolean
          default: true
          description: Process in parallel
    ScoringResult:
      type: object
      required: [entity_id, entity_type, engine_name, score, severity, timestamp]
      properties:
        entity_id:
          type: string
          description: Entity ID
        entity_type:
          $ref: '#/components/schemas/EntityType'
          description: Entity type
        engine_name:
          type: string
          description: Scoring engine name
        score:
          type: number
          format: float
          description: Risk score (0-100)
        severity:
          $ref: '#/components/schemas/SeverityLevel'
          description: Risk severity
        factors:
          type: object
          description: Score factors
        recommendations:
          type: array
          items:
            type: string
          description: Recommendations
        timestamp:
          type: string
          format: date-time
          description: Scoring timestamp
    HistoryRequest:
      type: object
      required: [date_from, date_to]
      properties:
        date_from:
          type: string
          format: date-time
          description: Start date
        date_to:
          type: string
          format: date-time
          description: End date
        engines:
          type: array
          items:
            type: string
          description: Engines to include
        granularity:
          type: string
          enum: [hourly, daily, weekly, monthly]
          description: Data granularity
        limit:
          type: integer
          description: Maximum results
    EnginesList:
      type: object
      required: [engines, total, available]
      properties:
        engines:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/EngineInfo'
          description: Engine information
        total:
          type: integer
          description: Total engines
        available:
          type: integer
          description: Available engines
    EngineInfo:
      type: object
      required: [name, status, description]
      properties:
        name:
          type: string
          description: Engine name
        status:
          type: string
          enum: [available, unavailable, degraded]
          description: Engine status
        description:
          type: string
          description: Engine description
        version:
          type: string
          description: Engine version

    # Agent Schemas
...